Moved the route builder and added the setStream
