+++ src/Attributes/AttributeContract.php
@@ -1,7 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace willitscale\Streetlamp\Attributes;
 
 interface AttributeContract
 {
+    public function bind(string $key): void;
 }
