+++ src/Builders/RouteBuilder.php
@@ -2,13 +2,17 @@
 
 declare(strict_types=1);
 
-namespace willitscale\Streetlamp;
+namespace willitscale\Streetlamp\Builders;
 
-use willitscale\Streetlamp\Attributes\RouteContract;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use ReflectionClass;
+use ReflectionMethod;
+use ReflectionParameter;
 use willitscale\Streetlamp\Attributes\DataBindings\ArrayMapInterface;
 use willitscale\Streetlamp\Attributes\Parameter\Parameter;
+use willitscale\Streetlamp\Attributes\RouteContract;
 use willitscale\Streetlamp\Attributes\Validators\ValidatorInterface;
-use willitscale\Streetlamp\Builders\RouterConfigBuilder;
 use willitscale\Streetlamp\Exceptions\CacheFileDoesNotExistException;
 use willitscale\Streetlamp\Exceptions\CacheFileInvalidFormatException;
 use willitscale\Streetlamp\Exceptions\ComposerFileDoesNotExistException;
@@ -20,11 +24,6 @@ use willitscale\Streetlamp\Exceptions\StreetLampException;
 use willitscale\Streetlamp\Models\Controller;
 use willitscale\Streetlamp\Models\Route;
 use willitscale\Streetlamp\Models\RouterConfig;
-use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
-use ReflectionClass;
-use ReflectionMethod;
-use ReflectionParameter;
 
 readonly class RouteBuilder
 {
