+++ src/Responses/ResponseHandler.php
@@ -12,13 +12,13 @@ use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use Psr\Log\LoggerInterface;
 use Throwable;
+use willitscale\Streetlamp\Builders\RouteBuilder;
 use willitscale\Streetlamp\Exceptions\InvalidParameterTypeException;
 use willitscale\Streetlamp\Exceptions\InvalidRouteResponseException;
 use willitscale\Streetlamp\Exceptions\Json\InvalidJsonObjectParameter;
 use willitscale\Streetlamp\Exceptions\Validators\InvalidParameterFailedToPassFilterValidation;
 use willitscale\Streetlamp\Models\Route;
 use willitscale\Streetlamp\Requests\Stream;
-use willitscale\Streetlamp\RouteBuilder;
 
 readonly class ResponseHandler implements RequestHandlerInterface
 {
